Refactoring, split unit test into multiple ones
