+++ onnxmltools/convert/sklearn/operator_converters/__init__.py
@@ -7,6 +7,8 @@
 # To register converter for scikit-learn operators, import associated modules here.
 from . import Binarizer
 from . import DictVectorizer
+from . import DecisionTree
+from . import GradientBoosting
 from . import Imputer
 from . import KNN
 from . import LabelEncoder
@@ -15,7 +17,7 @@ from . import LinearRegressor
 from . import NaiveBayes
 from . import Normalizer
 from . import OneHotEncoder
+from . import RandomForest
 from . import Scaler
 from . import SVM
 from . import SVD
-from . import TreeEnsemble
