+++ onnxmltools/utils/__init__.py
@@ -14,3 +14,5 @@ from .metadata_props import add_metadata_props, set_denotation
 from .visualize import visualize_model
 from .float16_converter import convert_float_to_float16
 from .tests_helper import dump_data_and_model
+from .tests_helper import dump_one_class_classification, dump_binary_classification, dump_multiple_classification
+from .tests_helper import dump_multiple_regression, dump_single_regression
