+++ onnxmltools/utils/tests_helper.py
@@ -93,7 +93,7 @@ def dump_data_and_model(data, model, onnx=None, basename="model", folder="tests"
         array = numpy.array(data)
         if inputs is None:
             inputs = [('input', FloatTensorType(list(array.shape)))]
-        onnx = convert_model(model, basename, inputs)
+        onnx, _ = convert_model(model, basename, inputs)
     
     dest = os.path.join(folder, basename + ".model.onnx")
     names.append(dest)
@@ -111,16 +111,127 @@ def convert_model(model, name, input_types):
     :return: *onnx* model
     """
     from sklearn.base import BaseEstimator
-    if isinstance(model, BaseEstimator):
+    if model.__class__.__name__.startswith("LGBM"):
+        from onnxmltools.convert import convert_lightgbm
+        model, prefix = convert_lightgbm(model, name, input_types), "LightGbm"
+    elif isinstance(model, BaseEstimator):
         from onnxmltools.convert import convert_sklearn
-        return convert_sklearn(model, name, input_types)
+        model, prefix = convert_sklearn(model, name, input_types), "Sklearn"
     else:
         from keras.models import Model
         if isinstance(model, Model):
             from onnxmltools.convert import convert_keras
-            return convert_keras(model, name, input_types)
+            model, prefix = convert_keras(model, name, input_types), "Keras"
         else:
             from onnxmltools.convert import convert_coreml
-            return convert_coreml(model, name, input_types)
+            model, prefix = convert_coreml(model, name, input_types), "Cml"
+    if model is None:
+        raise RuntimeError("Unable to convert model of type '{0}'.".format(type(model)))
+    return model, prefix
     
     
+def dump_one_class_classification(model, suffix="", folder="tests"):
+    """
+    Trains and dumps a model for a One Class outlier problem.
+    The function trains a model and calls
+    :func:`dump_data_and_model`.
+    
+    :param model: any model following *scikit-learn* API
+    :param suffix: added to filenames
+    :param folder: where to save the file
+    :return: output of :func:`dump_data_and_model`
+    
+    Every created filename will follow the pattern:
+    ``<folder>/<prefix><task><classifier-name><suffix>.<data|expected|model|onnx>.<pkl|onnx>``.
+    """
+    X = [[0., 1.], [1., 1.], [2., 0.]]
+    X = numpy.array(X, dtype=numpy.float32)
+    y = [1, 1, 1]
+    model.fit(X, y)
+    model_onnx, prefix = convert_model(model, 'one_class', [('input', FloatTensorType([1, 2]))])
+    return dump_data_and_model(X, model, model_onnx, folder=folder,
+                               basename=prefix + "One" + model.__class__.__name__ + suffix)
+
+
+def dump_binary_classification(model, suffix="", folder="tests"):
+    """
+    Trains and dumps a model for a binary classification problem.
+    
+    :param model: any model following *scikit-learn* API
+    :param suffix: added to filenames
+    :param folder: where to save the file
+    :return: output of :func:`dump_data_and_model`
+    
+    Every created filename will follow the pattern:
+    ``<folder>/<prefix><task><classifier-name><suffix>.<data|expected|model|onnx>.<pkl|onnx>``.
+    """
+    X = [[0, 1], [1, 1], [2, 0]]
+    X = numpy.array(X, dtype=numpy.float32)
+    y = ['A', 'B', 'A']
+    model.fit(X, y)
+    model_onnx, prefix = convert_model(model, 'tree-based binary classifier', [('input', FloatTensorType([1, 2]))])
+    dump_data_and_model(X, model, model_onnx, folder=folder,
+                        basename=prefix + "Bin" + model.__class__.__name__ + suffix)
+
+def dump_multiple_classification(model, suffix="", folder="tests"):
+    """
+    Trains and dumps a model for a binary classification problem.
+    
+    :param model: any model following *scikit-learn* API
+    :param suffix: added to filenames
+    :param folder: where to save the file
+    :return: output of :func:`dump_data_and_model`
+    
+    Every created filename will follow the pattern:
+    ``<folder>/<prefix><task><classifier-name><suffix>.<data|expected|model|onnx>.<pkl|onnx>``.
+    """
+    X = [[0, 1], [1, 1], [2, 0], [0.5, 0.5], [1.1, 1.1], [2.1, 0.1]]
+    X = numpy.array(X, dtype=numpy.float32)
+    y = [0, 1, 2, 1, 1, 2]
+    model.fit(X, y)
+    model_onnx, prefix = convert_model(model, 'tree-based multi-output regressor', [('input', FloatTensorType([1, 2]))])
+    dump_data_and_model(X, model, model_onnx, folder=folder,
+                        basename=prefix + "Mcl" + model.__class__.__name__ + suffix)
+
+
+def dump_multiple_regression(model, suffix="", folder="tests"):
+    """
+    Trains and dumps a model for a multi regression problem.
+    
+    :param model: any model following *scikit-learn* API
+    :param suffix: added to filenames
+    :param folder: where to save the file
+    :return: output of :func:`dump_data_and_model`
+    
+    Every created filename will follow the pattern:
+    ``<folder>/<prefix><task><classifier-name><suffix>.<data|expected|model|onnx>.<pkl|onnx>``.
+    """
+    X = [[0, 1], [1, 1], [2, 0]]
+    X = numpy.array(X, dtype=numpy.float32)
+    y = numpy.array([[100, 50], [100, 49], [100, 99]], dtype=numpy.float32)
+    model.fit(X, y)
+    model_onnx, prefix = convert_model(model, 'tree-based multi-output regressor', [('input', FloatTensorType([1, 2]))])
+    dump_data_and_model(X, model, model_onnx, folder=folder,
+                        basename=prefix + "MRg" + model.__class__.__name__ + suffix)
+
+
+def dump_single_regression(model, suffix="", folder="tests"):
+    """
+    Trains and dumps a model for a regression problem.
+    
+    :param model: any model following *scikit-learn* API
+    :param prefix: library name
+    :param suffix: added to filenames
+    :param folder: where to save the file
+    :return: output of :func:`dump_data_and_model`
+    
+    Every created filename will follow the pattern:
+    ``<folder>/<prefix><task><classifier-name><suffix>.<data|expected|model|onnx>.<pkl|onnx>``.
+    """
+    X = [[0, 1], [1, 1], [2, 0]]
+    X = numpy.array(X, dtype=numpy.float32)
+    y = numpy.array([100, -10, 50], dtype=numpy.float32)
+    model.fit(X, y)
+    model_onnx, prefix = convert_model(model, 'tree-based regressor', [('input', FloatTensorType([1, 2]))])
+    dump_data_and_model(X, model, model_onnx, folder=folder,
+                        basename=prefix + "Reg" + model.__class__.__name__ + suffix)
